fix: 更新代码质量限制并添加审计助手功能
